+++ discord_announcer/urls.py
@@ -10,4 +10,6 @@ app_name: str = "discord_announcer"
 
 urlpatterns = [
     path("", views.index, name="index"),
+    path("check-tokens/", views.check_tokens, name="check_tokens"),
+    path("send-latest/", views.send_latest, name="send_latest"),
 ]